Move race functions to RHRace
