+++ src/server/RHRace.py
@@ -4,16 +4,24 @@ import logging
 import json
 import RHUtils
 import Results
+import gevent
+import Config
+import random
+from datetime import datetime
+from flask import request
 from monotonic import monotonic
 from eventmanager import Evt
+from util.InvokeFuncQueue import InvokeFuncQueue
+from RHUtils import catchLogExceptionsWrapper
+from led_event_manager import ColorVal
 
 logger = logging.getLogger(__name__)
 
 class RHRace():
     '''Class to hold race management variables.'''
-    def __init__(self, RaceContext):
+    def __init__(self, racecontext):
         # internal references
-        self._racecontext = RaceContext
+        self._racecontext = racecontext
         self.__ = self._racecontext.language.__
         # setup/options
         self.num_nodes = 0
@@ -60,6 +68,7 @@ class RHRace():
 
         self._seat_colors = []
         self.external_flag = False # is race data controlled externally (from cluster)
+        self.pass_invoke_func_queue_obj = InvokeFuncQueue(logger)
 
         self.clear_results()
 
@@ -72,6 +81,867 @@ class RHRace():
             source
             deleted
         '''
+
+    @catchLogExceptionsWrapper
+    def stage(self, data=None):
+        if data and data.get('secondary_format'):
+            self.format = SECONDARY_RACE_FORMAT
+    
+        assigned_start = data.get('start_time_s', False) if data else None
+    
+        heat_data = self._racecontext.rhdata.get_heat(self.current_heat)
+        race_format = self.format
+        if race_format is SECONDARY_RACE_FORMAT:  # if running as secondary timer
+            check_create_sec_format_heat()
+        
+        self._racecontext.rhdata.clear_lapSplits()  # clear lap-splits from previous race
+    
+        heat_data = self._racecontext.rhdata.get_heat(self.current_heat)
+    
+        if heat_data:
+            heatNodes = self._racecontext.rhdata.get_heatNodes_by_heat(self.current_heat)
+            pilot_names_list = []
+            for heatNode in heatNodes:
+                if heatNode.node_index is not None and heatNode.node_index < self.num_nodes:
+                    if heatNode.pilot_id != RHUtils.PILOT_ID_NONE:
+                        pilot_obj = self._racecontext.rhdata.get_pilot(heatNode.pilot_id)
+                        if pilot_obj and pilot_obj.callsign:
+                            pilot_names_list.append(pilot_obj.callsign)
+    
+            if request and len(pilot_names_list) <= 0:
+                self._racecontext.rhui.emit_priority_message(self._racecontext.__('No valid pilots in race'), True, nobroadcast=True)
+    
+            logger.info("Staging new race, format: {}".format(getattr(race_format, "name", "????")))
+            max_round = self._racecontext.rhdata.get_max_round(self.current_heat)
+            if max_round is None:
+                max_round = 0
+            logger.info("Racing heat '{}' round {}, pilots: {}".format(heat_data.display_name, (max_round+1),
+                                                                       ", ".join(pilot_names_list)))
+        else:
+            heatNodes = []
+    
+            profile_freqs = json.loads(self.profile.frequencies)
+    
+            class FauxHeatNode():
+                node_index = None
+                pilot_id = 1
+    
+            for idx in range(self.num_nodes):
+                if (profile_freqs["f"][idx]):
+                    heatNode = FauxHeatNode
+                    heatNode.node_index = idx
+                    heatNodes.append(heatNode)
+    
+        if self._racecontext.cluster:
+            self._racecontext.cluster.emitToSplits('stage_race')
+    
+        if self.race_status != RaceStatus.READY:
+            if race_format is SECONDARY_RACE_FORMAT:  # if running as secondary timer
+                if self.race_status == RaceStatus.RACING:
+                    return  # if race in progress then leave it be
+                # if missed stop/discard message then clear current race
+                logger.info("Forcing race clear/restart because running as secondary timer")
+                self.discard_laps()
+            elif self.race_status == RaceStatus.DONE and not self.any_laps_recorded():
+                self.on_discard_laps()  # if no laps then allow restart
+
+        if self.race_status == RaceStatus.READY: # only initiate staging if ready
+            # common race start events (do early to prevent processing delay when start is called)
+            self._racecontext.interface.enable_calibration_mode() # Nodes reset triggers on next pass
+    
+            if race_format is not SECONDARY_RACE_FORMAT: # don't enforce class format if running as secondary timer
+                if heat_data and heat_data.class_id != RHUtils.CLASS_ID_NONE:
+                    class_format_id = self._racecontext.rhdata.get_raceClass(heat_data.class_id).format_id
+                    if class_format_id != RHUtils.FORMAT_ID_NONE:
+                        self.format = self._racecontext.rhdata.get_raceFormat(class_format_id)
+                        self._racecontext.rhui.emit_current_laps()
+                        logger.info("Forcing race format from class setting: '{0}' ({1})".format(self.format.name, self.format.id))
+    
+            self.clear_laps() # Clear laps before race start
+            init_node_cross_fields()  # set 'cur_pilot_id' and 'cross' fields on nodes
+            self._racecontext.last_race = None # clear all previous race data
+            self.timer_running = False # indicate race timer not running
+            self.race_status = RaceStatus.STAGING
+            self.win_status = WinStatus.NONE
+            self.race_winner_name = ''
+            self.race_winner_phonetic = ''
+            self.race_winner_lap_id = 0
+            self.race_winner_pilot_id = RHUtils.PILOT_ID_NONE
+            self.race_leader_lap = 0  # clear current race leader
+            self.race_leader_pilot_id = RHUtils.PILOT_ID_NONE
+            self.status_message = ''
+            self.any_races_started = True
+    
+            self.init_node_finished_flags(heatNodes)
+    
+            self._racecontext.interface.set_race_status(RaceStatus.STAGING)
+            self._racecontext.rhui.emit_current_laps() # Race page, blank laps to the web client
+            self._racecontext.rhui.emit_current_leaderboard() # Race page, blank leaderboard to the web client
+            self._racecontext.rhui.emit_race_status()
+    
+            assigned_start_ok_flag = False
+            if assigned_start:
+                self.stage_time_monotonic = monotonic() + float(Config.GENERAL['RACE_START_DELAY_EXTRA_SECS'])
+                if assigned_start > self.stage_time_monotonic:
+                    staging_tones = 0
+                    hide_stage_timer = True
+                    self.start_time_monotonic = assigned_start
+                    assigned_start_ok_flag = True
+    
+            if not assigned_start_ok_flag:
+                staging_fixed_ms = (0 if race_format.staging_fixed_tones <= 1 else race_format.staging_fixed_tones - 1) * 1000
+    
+                staging_random_ms = random.randint(0, race_format.start_delay_max_ms)
+                hide_stage_timer = (race_format.start_delay_max_ms > 0)
+    
+                staging_total_ms = staging_fixed_ms + race_format.start_delay_min_ms + staging_random_ms
+    
+                if race_format.staging_delay_tones == StagingTones.TONES_NONE:
+                    if staging_total_ms > 0:
+                        staging_tones = race_format.staging_fixed_tones
+                    else:
+                        staging_tones = staging_fixed_ms / 1000
+                else:
+                    staging_tones = staging_total_ms // 1000
+                    if staging_random_ms % 1000:
+                        staging_tones += 1
+    
+                self.stage_time_monotonic = monotonic() + float(Config.GENERAL['RACE_START_DELAY_EXTRA_SECS'])
+                self.start_time_monotonic = self.stage_time_monotonic + (staging_total_ms / 1000 )
+    
+            self.start_time_epoch_ms = monotonic_to_epoch_millis(self.start_time_monotonic)
+            self.start_token = random.random()
+            gevent.spawn(self.race_start_thread, self.start_token)
+    
+            eventPayload = {
+                'hide_stage_timer': hide_stage_timer,
+                'pi_staging_at_s': self.stage_time_monotonic,
+                'staging_tones': staging_tones,
+                'pi_starts_at_s': self.start_time_monotonic,
+                'color': ColorVal.ORANGE,
+                'heat_id': self.current_heat,
+            }
+    
+            eventPayload['race_node_colors'] = self.seat_colors
+    
+            self._racecontext.events.trigger(Evt.RACE_STAGE, eventPayload)
+    
+            SOCKET_IO.emit('stage_ready', {
+                'hide_stage_timer': hide_stage_timer,
+                'pi_staging_at_s': self.stage_time_monotonic,
+                'staging_tones': staging_tones,
+                'pi_starts_at_s': self.start_time_monotonic,
+                'unlimited_time': race_format.unlimited_time,
+                'race_time_sec': race_format.race_time_sec,
+            }) # Announce staging with final parameters
+    
+        else:
+            logger.info("Attempted to stage race while status is not 'ready'")
+
+
+    @catchLogExceptionsWrapper
+    def race_start_thread(self, start_token):
+        # clear any lingering crossings at staging (if node rssi < enterAt)
+        for node in self._racecontext.interface.nodes:
+            if node.crossing_flag and node.frequency > 0 and \
+                (self.format is SECONDARY_RACE_FORMAT or
+                (node.current_pilot_id != RHUtils.PILOT_ID_NONE and node.current_rssi < node.enter_at_level)):
+                logger.info("Forcing end crossing for node {0} at staging (rssi={1}, enterAt={2}, exitAt={3})".\
+                           format(node.index+1, node.current_rssi, node.enter_at_level, node.exit_at_level))
+                self._racecontext.interface.force_end_crossing(node.index)
+    
+        if self._racecontext.cluster and self._racecontext.cluster.hasSecondaries():
+            self._racecontext.cluster.doClusterRaceStart()
+    
+        # set lower EnterAt/ExitAt values if configured
+        if self._racecontext.rhdata.get_optionInt('startThreshLowerAmount') > 0 and self._racecontext.rhdata.get_optionInt('startThreshLowerDuration') > 0:
+            lower_amount = self._racecontext.rhdata.get_optionInt('startThreshLowerAmount')
+            logger.info("Lowering EnterAt/ExitAt values at start of race, amount={0}%, duration={1} secs".\
+                        format(lower_amount, self._racecontext.rhdata.get_optionInt('startThreshLowerDuration')))
+            lower_end_time = self.start_time_monotonic + self._racecontext.rhdata.get_optionInt('startThreshLowerDuration')
+            for node in self._racecontext.interface.nodes:
+                if node.frequency > 0 and (self.format is SECONDARY_RACE_FORMAT or node.current_pilot_id != RHUtils.PILOT_ID_NONE):
+                    if node.current_rssi < node.enter_at_level:
+                        diff_val = int((node.enter_at_level-node.exit_at_level)*lower_amount/100)
+                        if diff_val > 0:
+                            new_enter_at = node.enter_at_level - diff_val
+                            new_exit_at = max(node.exit_at_level - diff_val, 0)
+                            if node.api_valid_flag and node.is_valid_rssi(new_enter_at):
+                                logger.info("For node {0} lowering EnterAt from {1} to {2} and ExitAt from {3} to {4}"\
+                                        .format(node.index+1, node.enter_at_level, new_enter_at, node.exit_at_level, new_exit_at))
+                                node.start_thresh_lower_time = lower_end_time  # set time when values will be restored
+                                node.start_thresh_lower_flag = True
+                                # use 'transmit_' instead of 'set_' so values are not saved in node object
+                                self._racecontext.interface.transmit_enter_at_level(node, new_enter_at)
+                                self._racecontext.interface.transmit_exit_at_level(node, new_exit_at)
+                        else:
+                            logger.info("Not lowering EnterAt/ExitAt values for node {0} because EnterAt value ({1}) unchanged"\
+                                    .format(node.index+1, node.enter_at_level))
+                    else:
+                        logger.info("Not lowering EnterAt/ExitAt values for node {0} because current RSSI ({1}) >= EnterAt ({2})"\
+                                .format(node.index+1, node.current_rssi, node.enter_at_level))
+    
+        # do non-blocking delay before time-critical code
+        while (monotonic() < self.start_time_monotonic - 0.5):
+            gevent.sleep(0.1)
+    
+        if self.race_status == RaceStatus.STAGING and \
+            self.start_token == start_token:
+            # Only start a race if it is not already in progress
+            # Null this thread if token has changed (race stopped/started quickly)
+    
+            # do blocking delay until race start
+            while monotonic() < self.start_time_monotonic:
+                pass
+    
+            # !!! RACE STARTS NOW !!!
+    
+            # do time-critical tasks
+            self._racecontext.events.trigger(Evt.RACE_START, {
+                'heat_id': self.current_heat,
+                'color': ColorVal.GREEN
+                })
+    
+            # do secondary start tasks (small delay is acceptable)
+            self.start_time = datetime.now() # record standard-formatted time
+    
+            for node in self._racecontext.interface.nodes:
+                node.history_values = [] # clear race history
+                node.history_times = []
+                node.under_min_lap_count = 0
+                # clear any lingering crossing (if rssi>enterAt then first crossing starts now)
+                if node.crossing_flag and node.frequency > 0 and (
+                    self.format is SECONDARY_RACE_FORMAT or node.current_pilot_id != RHUtils.PILOT_ID_NONE):
+                    logger.info("Forcing end crossing for node {0} at start (rssi={1}, enterAt={2}, exitAt={3})".\
+                               format(node.index+1, node.current_rssi, node.enter_at_level, node.exit_at_level))
+                    self._racecontext.interface.force_end_crossing(node.index)
+    
+            self.race_status = RaceStatus.RACING # To enable registering passed laps
+            self._racecontext.interface.set_race_status(RaceStatus.RACING)
+            self.timer_running = True # indicate race timer is running
+    
+            # kick off race expire processing
+            race_format = self.format
+            if race_format and race_format.unlimited_time == 0: # count down
+                gevent.spawn(self.race_expire_thread, start_token)
+    
+            self._racecontext.rhui.emit_race_status() # Race page, to set race button states
+            logger.info('Race started at {:.3f} ({:.0f})'.format(self.start_time_monotonic, self.start_time_epoch_ms))
+    
+    @catchLogExceptionsWrapper
+    def race_expire_thread(self, start_token):
+        race_format = self.format
+        if race_format and race_format.unlimited_time == 0: # count down
+            gevent.sleep(race_format.race_time_sec)
+            # if race still in progress and is still same race
+            if self.race_status == RaceStatus.RACING and self.start_token == start_token:
+                logger.info("Race count-down timer reached expiration")
+                self.timer_running = False # indicate race timer no longer running
+                self._racecontext.events.trigger(Evt.RACE_FINISH, {
+                    'heat_id': self.current_heat,
+                    })
+                self.pass_invoke_func_queue_obj.waitForQueueEmpty()  # wait until any active pass-record processing is finished
+                self.check_win_condition(at_finish=True, start_token=start_token)
+                self._racecontext.rhui.emit_current_leaderboard()
+                if race_format.lap_grace_sec > -1:
+                    gevent.sleep((self.start_time_monotonic + race_format.race_time_sec + race_format.lap_grace_sec) - monotonic())
+                    if self.race_status == RaceStatus.RACING and self.start_token == start_token:
+                        self.stop()
+                        logger.debug("Race grace period reached")
+                    else:
+                        logger.debug("Grace period timer {} is unused".format(start_token))
+            else:
+                logger.debug("Race-time-expire thread {} is unused".format(start_token))
+    
+    @catchLogExceptionsWrapper
+    def stop(self, doSave=False):
+        '''Stops the race and stops registering laps.'''
+        if self._racecontext.cluster:
+            self._racecontext.cluster.emitToSplits('stop_race')
+    
+        if self.race_status == RaceStatus.RACING:
+            # clear any crossings still in progress
+            any_forced_flag = False
+            for node in self._racecontext.interface.nodes:
+                if node.crossing_flag and node.frequency > 0 and \
+                                node.current_pilot_id != RHUtils.PILOT_ID_NONE:
+                    logger.info("Forcing end crossing for node {} at race stop (rssi={}, enterAt={}, exitAt={})".\
+                                format(node.index+1, node.current_rssi, node.enter_at_level, node.exit_at_level))
+                    self._racecontext.interface.force_end_crossing(node.index)
+                    any_forced_flag = True
+            if any_forced_flag:  # give forced end-crossings a chance to complete before stopping race
+                gevent.spawn_later(0.5, self.do_stop_race_actions, doSave)
+            else:
+                self.do_stop_race_actions(doSave)
+        else:
+            self.do_stop_race_actions(doSave)
+    
+        SOCKET_IO.emit('stop_timer') # Loop back to race page to stop the timer
+    
+    @catchLogExceptionsWrapper
+    def do_stop_race_actions(self, doSave=False):
+        if self.race_status == RaceStatus.RACING:
+            self.end_time = monotonic() # Update the race end time stamp
+            delta_time = self.end_time - self.start_time_monotonic
+    
+            logger.info('Race stopped at {:.3f} ({:.0f}), duration {:.0f}s'.format(self.end_time, monotonic_to_epoch_millis(self.end_time), delta_time))
+    
+            min_laps_list = []  # show nodes with laps under minimum (if any)
+            for node in self._racecontext.interface.nodes:
+                if node.under_min_lap_count > 0:
+                    min_laps_list.append('Node {0} Count={1}'.format(node.index+1, node.under_min_lap_count))
+            if len(min_laps_list) > 0:
+                logger.info('Nodes with laps under minimum:  ' + ', '.join(min_laps_list))
+    
+            self.race_status = RaceStatus.DONE # To stop registering passed laps, waiting for laps to be cleared
+            self._racecontext.interface.set_race_status(RaceStatus.DONE)
+    
+            self._racecontext.events.trigger(Evt.RACE_STOP, {
+                'heat_id': self.current_heat,
+                'color': ColorVal.RED
+            })
+            self.pass_invoke_func_queue_obj.waitForQueueEmpty()  # wait until any active pass-record processing is finished
+            self.check_win_condition()
+    
+            if self._racecontext.cluster and self._racecontext.cluster.hasSecondaries():
+                self._racecontext.cluster.doClusterRaceStop()
+    
+        elif self.race_status == RaceStatus.STAGING:
+            logger.info('Stopping race during staging')
+            self.race_status = RaceStatus.READY # Go back to ready state
+            self._racecontext.interface.set_race_status(RaceStatus.READY)
+            self._racecontext.events.trigger(Evt.LAPS_CLEAR)
+            delta_time = 0
+    
+        else:
+            self.race_status = RaceStatus.DONE # To stop registering passed laps, waiting for laps to be cleared
+            self._racecontext.interface.set_race_status(RaceStatus.DONE)
+    
+            logger.debug('No active race to stop')
+            delta_time = 0
+    
+        # check if nodes may be set to temporary lower EnterAt/ExitAt values (and still have them)
+        if self._racecontext.rhdata.get_optionInt('startThreshLowerAmount') > 0 and \
+                delta_time < self._racecontext.rhdata.get_optionInt('startThreshLowerDuration'):
+            for node in self._racecontext.interface.nodes:
+                # if node EnterAt/ExitAt values need to be restored then do it soon
+                if node.frequency > 0 and (
+                    self.format is SECONDARY_RACE_FORMAT or (
+                        node.current_pilot_id != RHUtils.PILOT_ID_NONE and \
+                        node.start_thresh_lower_flag)):
+                    node.start_thresh_lower_time = self.end_time + 0.1
+    
+        self.timer_running = False # indicate race timer not running
+        self.scheduled = False # also stop any deferred start
+    
+        self._racecontext.rhui.emit_race_status() # Race page, to set race button states
+        self._racecontext.rhui.emit_current_leaderboard()
+    
+        if doSave:
+            self.do_save_actions()
+
+    @catchLogExceptionsWrapper
+    def save(self):
+        '''Handle "save" UI action'''
+    
+        if self.race_status == RaceStatus.RACING:
+            self.stop(doSave=True)
+        else:
+            self.do_save_actions()
+    
+    @catchLogExceptionsWrapper
+    def do_save_actions(self):
+        '''Save current laps data to the database.'''
+        if self.current_heat == RHUtils.HEAT_ID_NONE:
+            self.on_discard_laps(saved=True)
+            return False
+    
+        if self._racecontext.cluster:
+            self._racecontext.cluster.emitToSplits('save_laps')
+    
+        heat = self._racecontext.rhdata.get_heat(self.current_heat)
+    
+        # Clear caches
+        self._racecontext.rhdata.clear_results_heat(self.current_heat)
+        self._racecontext.rhdata.clear_results_raceClass(heat.class_id)
+        self._racecontext.rhdata.clear_results_event()
+    
+        # Get the last saved round for the current heat
+        max_round = self._racecontext.rhdata.get_max_round(self.current_heat)
+    
+        if max_round is None:
+            max_round = 0
+        # Loop through laps to copy to saved races
+        profile = self.profile
+        profile_freqs = json.loads(profile.frequencies)
+    
+        new_race_data = {
+            'round_id': max_round+1,
+            'heat_id': self.current_heat,
+            'class_id': heat.class_id,
+            'format_id': self.format.id if hasattr(self.format, 'id') else RHUtils.FORMAT_ID_NONE,
+            'start_time': self.start_time_monotonic,
+            'start_time_formatted': self.start_time.strftime("%Y-%m-%d %H:%M:%S"),
+            }
+    
+        new_race = self._racecontext.rhdata.add_savedRaceMeta(new_race_data)
+    
+        race_data = {}
+    
+        for node_index in range(self.num_nodes):
+            if profile_freqs["f"][node_index] != RHUtils.FREQUENCY_ID_NONE:
+                pilot_id = self._racecontext.rhdata.get_pilot_from_heatNode(self.current_heat, node_index)
+    
+                if pilot_id is not None:
+                    race_data[node_index] = {
+                        'race_id': new_race.id,
+                        'pilot_id': pilot_id,
+                        'history_values': json.dumps(self._racecontext.interface.nodes[node_index].history_values),
+                        'history_times': json.dumps(self._racecontext.interface.nodes[node_index].history_times),
+                        'enter_at': self._racecontext.interface.nodes[node_index].enter_at_level,
+                        'exit_at': self._racecontext.interface.nodes[node_index].exit_at_level,
+                        'frequency': self._racecontext.interface.nodes[node_index].frequency,
+                        'laps': self.node_laps[node_index]
+                        }
+    
+                    self._racecontext.rhdata.set_pilot_used_frequency(pilot_id, {
+                        'b': profile_freqs["b"][node_index],
+                        'c': profile_freqs["c"][node_index],
+                        'f': profile_freqs["f"][node_index]
+                        })
+    
+        self._racecontext.rhdata.add_race_data(race_data)
+    
+        self._racecontext.events.trigger(Evt.LAPS_SAVE, {
+            'race_id': new_race.id,
+            })
+    
+        logger.info('Current laps saved: Heat {0} Round {1}'.format(self.current_heat, max_round+1))
+    
+        self.on_discard_laps(saved=True) # Also clear the current laps
+    
+        next_heat = self._racecontext.rhdata.get_next_heat_id(heat)
+        if next_heat is not heat.id:
+            on_set_current_heat({'heat': next_heat})
+    
+        # spawn thread for updating results caches
+        cache_params = {
+            'race_id': new_race.id,
+            'heat_id': new_race.heat_id,
+            'round_id': new_race.round_id,
+        }
+        gevent.spawn(build_atomic_result_caches, cache_params)
+        
+        self._racecontext.rhui.emit_race_saved(new_race, race_data)
+
+    def add_lap(self, node, lap_timestamp_absolute, source):
+        '''Handles pass records from the nodes.'''
+    
+        logger.debug('Pass record: Node={}, abs_ts={:.3f}, source={} ("{}")' \
+                     .format(node.index+1, lap_timestamp_absolute, source, self._racecontext.interface.get_lap_source_str(source)))
+        node.pass_crossing_flag = False  # clear the "synchronized" version of the crossing flag
+        node.debug_pass_count += 1
+        self._racecontext.rhui.emit_node_data() # For updated triggers and peaks
+    
+        profile_freqs = json.loads(self.profile.frequencies)
+        if profile_freqs["f"][node.index] != RHUtils.FREQUENCY_ID_NONE :
+            # always count laps if race is running, otherwise test if lap should have counted before race end
+            if self.race_status is RaceStatus.RACING \
+                or (self.race_status is RaceStatus.DONE and \
+                    lap_timestamp_absolute < self.end_time):
+    
+                # Get the current pilot id on the node
+                pilot_id = self._racecontext.rhdata.get_pilot_from_heatNode(self.current_heat, node.index)
+    
+                # reject passes before race start and with disabled (no-pilot) nodes
+                race_format = self.format
+                if (pilot_id is not None and pilot_id != RHUtils.PILOT_ID_NONE) or race_format is SECONDARY_RACE_FORMAT or self.current_heat is RHUtils.HEAT_ID_NONE:
+                    if lap_timestamp_absolute >= self.start_time_monotonic:
+    
+                        # if node EnterAt/ExitAt values need to be restored then do it soon
+                        if node.start_thresh_lower_flag:
+                            node.start_thresh_lower_time = monotonic()
+    
+                        lap_time_stamp = (lap_timestamp_absolute - self.start_time_monotonic)
+                        lap_time_stamp *= 1000 # store as milliseconds
+    
+                        lap_number = len(self.get_active_laps()[node.index])
+    
+                        if lap_number: # This is a normal completed lap
+                            # Find the time stamp of the last lap completed (including "late" laps for timing)
+                            last_lap_time_stamp = self.get_active_laps(True)[node.index][-1]['lap_time_stamp']
+    
+                            # New lap time is the difference between the current time stamp and the last
+                            lap_time = lap_time_stamp - last_lap_time_stamp
+    
+                        else: # No previous laps, this is the first pass
+                            # Lap zero represents the time from the launch pad to flying through the gate
+                            lap_time = lap_time_stamp
+                            node.first_cross_flag = True  # indicate first crossing completed
+    
+                        if race_format is SECONDARY_RACE_FORMAT:
+                            min_lap = 0  # don't enforce min-lap time if running as secondary timer
+                            min_lap_behavior = 0
+                        else:
+                            min_lap = self._racecontext.rhdata.get_optionInt("MinLapSec")
+                            min_lap_behavior = self._racecontext.rhdata.get_optionInt("MinLapBehavior")
+    
+                        lap_time_fmtstr = RHUtils.time_format(lap_time, self._racecontext.rhdata.get_option('timeFormat'))
+                        lap_ts_fmtstr = RHUtils.time_format(lap_time_stamp, self._racecontext.rhdata.get_option('timeFormat'))
+                        pilot_obj = self._racecontext.rhdata.get_pilot(pilot_id)
+                        pilot_namestr = pilot_obj.callsign if pilot_obj else ""
+    
+                        lap_ok_flag = True
+                        lap_late_flag = False
+                        pilot_done_flag = False
+                        if lap_number != 0:  # if initial lap then always accept and don't check lap time; else:
+                            if lap_time < (min_lap * 1000):  # if lap time less than minimum
+                                node.under_min_lap_count += 1
+                                logger.info('Pass record under lap minimum ({}): Node={}, lap={}, lapTime={}, sinceStart={}, count={}, source={}, pilot: {}' \
+                                           .format(min_lap, node.index+1, lap_number, \
+                                                   lap_time_fmtstr, lap_ts_fmtstr, \
+                                                   node.under_min_lap_count, self._racecontext.interface.get_lap_source_str(source), \
+                                                   pilot_namestr))
+                                if min_lap_behavior != 0:  # if behavior is 'Discard New Short Laps'
+                                    lap_ok_flag = False
+    
+                            if race_format.unlimited_time == 0 and \
+                                race_format.lap_grace_sec > -1 and \
+                                lap_time_stamp > (race_format.race_time_sec + race_format.lap_grace_sec)*1000:
+                                logger.info('Ignoring lap after grace period expired: Node={}, lap={}, lapTime={}, sinceStart={}, source={}, pilot: {}' \
+                                           .format(node.index+1, lap_number, lap_time_fmtstr, lap_ts_fmtstr, \
+                                                   self._racecontext.interface.get_lap_source_str(source), pilot_namestr))
+                                lap_ok_flag = False
+    
+                        if lap_ok_flag:
+                            node_finished_flag = self.get_node_finished_flag(node.index)
+                            # set next node race status as 'finished' if timer mode is count-down race and race-time has expired
+                            if (race_format.unlimited_time == 0 and lap_time_stamp > race_format.race_time_sec * 1000) or \
+                                (self.format.win_condition == WinCondition.FIRST_TO_LAP_X and lap_number >= race_format.number_laps_win):
+                                self.set_node_finished_flag(node.index)
+                                if not node_finished_flag:
+                                    logger.info('Pilot {} done'.format(pilot_obj.callsign if pilot_obj else node.index))
+                                    pilot_done_flag = True
+    
+                            if node_finished_flag:
+                                lap_late_flag = True  # "late" lap pass (after grace lap)
+                                logger.info('Ignoring lap after pilot done: Node={}, lap={}, lapTime={}, sinceStart={}, source={}, pilot: {}' \
+                                           .format(node.index+1, lap_number, lap_time_fmtstr, lap_ts_fmtstr, \
+                                                   self._racecontext.interface.get_lap_source_str(source), pilot_namestr))
+                                
+                            if self.win_status == WinStatus.DECLARED and \
+                                race_format.unlimited_time == 1 and \
+                                self.format.team_racing_mode and \
+                                self.format.win_condition == WinCondition.FIRST_TO_LAP_X:
+                                lap_late_flag = True  # "late" lap pass after team race winner declared (when no time limit)
+                                if pilot_obj:
+                                    t_str = ", Team " + pilot_obj.team
+                                else:
+                                    t_str = ""
+                                logger.info('Ignoring lap after race winner declared: Node={}, lap={}, lapTime={}, sinceStart={}, source={}, pilot: {}{}' \
+                                           .format(node.index+1, lap_number, lap_time_fmtstr, lap_ts_fmtstr, \
+                                                   self._racecontext.interface.get_lap_source_str(source), pilot_namestr, t_str))
+    
+                            if logger.getEffectiveLevel() <= logging.DEBUG:  # if DEBUG msgs actually being logged
+                                late_str = " (late lap)" if lap_late_flag else ""
+                                enter_fmtstr = RHUtils.time_format((node.enter_at_timestamp-self.start_time_monotonic)*1000, \
+                                                                   self._racecontext.rhdata.get_option('timeFormat')) \
+                                               if node.enter_at_timestamp else "0"
+                                exit_fmtstr = RHUtils.time_format((node.exit_at_timestamp-self.start_time_monotonic)*1000, \
+                                                                  self._racecontext.rhdata.get_option('timeFormat')) \
+                                               if node.exit_at_timestamp else "0"
+                                logger.debug('Lap pass{}: Node={}, lap={}, lapTime={}, sinceStart={}, abs_ts={:.3f}, source={}, enter={}, exit={}, dur={:.0f}ms, pilot: {}' \
+                                            .format(late_str, node.index+1, lap_number, lap_time_fmtstr, lap_ts_fmtstr, \
+                                                    lap_timestamp_absolute, self._racecontext.interface.get_lap_source_str(source), \
+                                                    enter_fmtstr, exit_fmtstr, \
+                                                    (node.exit_at_timestamp-node.enter_at_timestamp)*1000, pilot_namestr))
+    
+                            # emit 'pass_record' message (to primary timer in cluster, livetime, etc).
+                            self._racecontext.rhui.emit_pass_record(node, lap_time_stamp)
+    
+                            # Add the new lap to the database
+                            lap_data = {
+                                'lap_number': lap_number,
+                                'lap_time_stamp': lap_time_stamp,
+                                'lap_time': lap_time,
+                                'lap_time_formatted': lap_time_fmtstr,
+                                'source': source,
+                                'deleted': lap_late_flag,  # delete if lap pass is after race winner declared
+                                'late_lap': lap_late_flag
+                            }
+                            self.node_laps[node.index].append(lap_data)
+    
+                            self.clear_results()
+    
+                            self._racecontext.events.trigger(Evt.RACE_LAP_RECORDED, {
+                                'pilot_id': pilot_id,
+                                'node_index': node.index,
+                                'peak_rssi': node.pass_peak_rssi,
+                                'frequency': node.frequency,
+                                'color': self.seat_colors[node.index],
+                                'lap': lap_data,
+                                'results': self.get_results(),
+                                'gap_info': Results.get_gap_info(self._racecontext, node.index),
+                                'pilot_done_flag': pilot_done_flag,
+                                })
+    
+                            if pilot_done_flag:
+                                self._racecontext.events.trigger(Evt.RACE_PILOT_DONE, {
+                                    'pilot_id': pilot_id,
+                                    'node_index': node.index,
+                                    'color': self.seat_colors[node.index],
+                                    'results': self.get_results(),
+                                    })
+    
+                            self._racecontext.rhui.emit_current_laps() # update all laps on the race page
+                            self._racecontext.rhui.emit_current_leaderboard() # generate and update leaderboard
+    
+                            if lap_number == 0:
+                                self._racecontext.rhui.emit_first_pass_registered(node.index) # play first-pass sound
+    
+                            if race_format.start_behavior == StartBehavior.FIRST_LAP:
+                                lap_number += 1
+    
+                            # announce lap
+                            if lap_number > 0:
+                                check_leader = race_format.win_condition != WinCondition.NONE and \
+                                               self.win_status != WinStatus.DECLARED
+                                # announce pilot lap number unless winner declared and pilot has finished final lap
+                                lap_id = lap_number if self.win_status != WinStatus.DECLARED or \
+                                                       (not node_finished_flag) else None
+                                if race_format.team_racing_mode:
+                                    team_name = pilot_obj.team if pilot_obj else ""
+                                    team_laps = self.team_results['meta']['teams'][team_name]['laps']
+                                    if not lap_late_flag:
+                                        logger.debug('Lap pass: Node={}, lap={}, pilot={} -> Team {} lap {}' \
+                                              .format(node.index+1, lap_number, pilot_namestr, team_name, team_laps))
+                                    # if winning team has been declared then don't announce team lap number
+                                    if self.win_status == WinStatus.DECLARED:
+                                        team_laps = None
+                                    self._racecontext.rhui.emit_phonetic_data(pilot_id, lap_id, lap_time, team_name, team_laps, \
+                                                    (check_leader and \
+                                                     team_name == Results.get_leading_team_name(self.team_results)), \
+                                                    node_finished_flag, node.index)
+                                else:
+                                    if check_leader:
+                                        leader_pilot_id = Results.get_leading_pilot_id(self, self._racecontext.interface, True)
+                                    else:
+                                        leader_pilot_id = RHUtils.PILOT_ID_NONE
+                                    self._racecontext.rhui.emit_phonetic_data(pilot_id, lap_id, lap_time, None, None, \
+                                                    (pilot_id == leader_pilot_id), node_finished_flag, node.index)
+                                    if leader_pilot_id != RHUtils.PILOT_ID_NONE:
+                                        # if new leading pilot was not called out above (different pilot) then call out now
+                                        if leader_pilot_id != pilot_id:
+                                            self._racecontext.rhui.emit_phonetic_leader(leader_pilot_id)
+                                            leader_pilot_obj = self._racecontext.rhdata.get_pilot(leader_pilot_id)
+                                            if leader_pilot_obj:
+                                                logger.info('Pilot {} is leading'.format(leader_pilot_obj.callsign))
+                                        else:
+                                            logger.info('Pilot {} is leading'.format(pilot_namestr))
+    
+                                # check for and announce possible winner (but wait until pass-record processing(s) is finished)
+                                self.pass_invoke_func_queue_obj.put(self.check_win_condition, emit_leaderboard_on_win=True) 
+    
+                        else:
+                            # record lap as 'invalid'
+                            self.node_laps[node.index].append({
+                                'lap_number': lap_number,
+                                'lap_time_stamp': lap_time_stamp,
+                                'lap_time': lap_time,
+                                'lap_time_formatted': lap_time_fmtstr,
+                                'source': source,
+                                'deleted': True,
+                                'invalid': True
+                            })
+                    else:
+                        logger.debug('Pass record dismissed: Node {}, Race not started (abs_ts={:.3f}, source={})' \
+                            .format(node.index+1, lap_timestamp_absolute, self._racecontext.interface.get_lap_source_str(source)))
+                else:
+                    logger.debug('Pass record dismissed: Node {}, Pilot not defined (abs_ts={:.3f}, source={})' \
+                        .format(node.index+1, lap_timestamp_absolute, self._racecontext.interface.get_lap_source_str(source)))
+        else:
+            logger.debug('Pass record dismissed: Node {}, Frequency not defined (abs_ts={:.3f}, source={})' \
+                .format(node.index+1, lap_timestamp_absolute, self._racecontext.interface.get_lap_source_str(source)))
+
+    @catchLogExceptionsWrapper
+    def delete_lap(self, data):
+        '''Delete a false lap.'''
+    
+        node_index = data['node']
+        lap_index = data['lap_index']
+    
+        if node_index is None or lap_index is None:
+            logger.error("Bad parameter in 'on_delete_lap()':  node_index={0}, lap_index={1}".format(node_index, lap_index))
+            return
+    
+        self.node_laps[node_index][lap_index]['invalid'] = True
+    
+        time = self.node_laps[node_index][lap_index]['lap_time_stamp']
+    
+        race_format = self.format
+        self.set_node_finished_flag(node_index, False)
+        lap_number = 0
+        for lap in self.node_laps[node_index]:
+            lap['deleted'] = False
+            if self.get_node_finished_flag(node_index):
+                lap['late_lap'] = True
+                lap['deleted'] = True
+            else:
+                lap['late_lap'] = False
+    
+            if lap.get('invalid', False):
+                lap['lap_number'] = None
+                lap['deleted'] = True
+            else:
+                lap['lap_number'] = lap_number
+                if race_format.unlimited_time == 0 and lap['lap_time_stamp'] > (race_format.race_time_sec * 1000) or \
+                    (race_format.win_condition == WinCondition.FIRST_TO_LAP_X and lap_number >= race_format.number_laps_win):
+                    self.set_node_finished_flag(node_index)
+                lap_number += 1
+    
+        db_last = False
+        db_next = False
+        for lap in self.node_laps[node_index]:
+            if not lap.get('invalid', False) and ((not lap['deleted']) or lap.get('late_lap', False)):
+                if lap['lap_time_stamp'] < time:
+                    db_last = lap
+                if lap['lap_time_stamp'] > time:
+                    db_next = lap
+                    break
+    
+        if db_next and db_last:
+            db_next['lap_time'] = db_next['lap_time_stamp'] - db_last['lap_time_stamp']
+            db_next['lap_time_formatted'] = RHUtils.time_format(db_next['lap_time'], self._racecontext.rhdata.get_option('timeFormat'))
+        elif db_next:
+            db_next['lap_time'] = db_next['lap_time_stamp']
+            db_next['lap_time_formatted'] = RHUtils.time_format(db_next['lap_time'], self._racecontext.rhdata.get_option('timeFormat'))
+    
+        try:  # delete any split laps for deleted lap
+            lap_splits = self._racecontext.rhdata.get_lapSplits_by_lap(node_index, lap_number)
+            if lap_splits and len(lap_splits) > 0:
+                for lap_split in lap_splits:
+                    self._racecontext.rhdata.clear_lapSplit(lap_split)
+        except:
+            logger.exception("Error deleting split laps")
+    
+        self._racecontext.events.trigger(Evt.LAP_DELETE, {
+            'node_index': node_index,
+            })
+    
+        logger.info('Lap deleted: Node {0} LapIndex {1}'.format(node_index+1, lap_index))
+    
+        self.clear_results()
+        self.pass_invoke_func_queue_obj.waitForQueueEmpty()  # wait until any active pass-record processing is finished
+        self.get_results()  # update leaderboard before checking possible updated winner/leader
+        self.check_win_condition(deletedLap=True)  # handle possible change in win status
+    
+        # handle possible change in race leader
+        leader_pilot_id = Results.get_leading_pilot_id(self._racecontext.race, self._racecontext.interface, True)
+        if leader_pilot_id != RHUtils.PILOT_ID_NONE:
+            self._racecontext.rhui.emit_phonetic_leader(leader_pilot_id)
+            leader_pilot_obj = self._racecontext.rhdata.get_pilot(leader_pilot_id)
+            if leader_pilot_obj:
+                logger.info('Pilot {} is leading (after deleted lap)'.format(leader_pilot_obj.callsign))
+    
+        self._racecontext.rhui.emit_current_laps() # Race page, update web client
+        self._racecontext.rhui.emit_current_leaderboard() # Race page, update web client
+
+    @catchLogExceptionsWrapper
+    def restore_deleted_lap(self, data):
+        '''Restore a deleted (or "late") lap.'''
+    
+        node_index = data['node']
+        lap_index = data['lap_index']
+    
+        if node_index is None or lap_index is None:
+            logger.error("Bad parameter in 'on_restore_deleted_lap()':  node_index={0}, lap_index={1}".format(node_index, lap_index))
+            return
+    
+        lap_obj = self.node_laps[node_index][lap_index]
+    
+        lap_obj['deleted'] = False
+        lap_obj['late_lap'] = False
+    
+        lap_number = 0  # adjust lap numbers and times as needed
+        last_lap_ts = 0
+        for idx, lap in enumerate(self.node_laps[node_index]):
+            if not lap['deleted']:
+                if idx >= lap_index:
+                    lap['lap_number'] = lap_number
+                    lap['lap_time'] = lap['lap_time_stamp'] - last_lap_ts
+                    lap['lap_time_formatted'] = RHUtils.time_format(lap['lap_time'], self._racecontext.rhdata.get_option('timeFormat'))
+                last_lap_ts = lap['lap_time_stamp']
+                lap_number += 1
+    
+        self._racecontext.events.trigger(Evt.LAP_RESTORE_DELETED, {
+            'node_index': node_index,
+            })
+    
+        logger.info('Restored deleted lap: Node {0} LapIndex {1}'.format(node_index+1, lap_index))
+    
+        self.clear_results()
+        self.pass_invoke_func_queue_obj.waitForQueueEmpty()  # wait until any active pass-record processing is finished
+        self.get_results()  # update leaderboard before checking possible updated winner/leader
+        self.check_win_condition(deletedLap=True)  # handle possible change in win status
+    
+        # handle possible change in race leader
+        leader_pilot_id = Results.get_leading_pilot_id(self._racecontext.race, self._racecontext.interface, True)
+        if leader_pilot_id != RHUtils.PILOT_ID_NONE:
+            self._racecontext.rhui.emit_phonetic_leader(leader_pilot_id)
+            leader_pilot_obj = self._racecontext.rhdata.get_pilot(leader_pilot_id)
+            if leader_pilot_obj:
+                logger.info('Pilot {} is leading (after deleted lap)'.format(leader_pilot_obj.callsign))
+    
+        self._racecontext.rhui.emit_current_laps() # Race page, update web client
+        self._racecontext.rhui.emit_current_leaderboard() # Race page, update web client
+
+    @catchLogExceptionsWrapper
+    def discard_laps(self, **kwargs):
+        '''Clear the current laps without saving.'''
+    
+        if self.race_status == RaceStatus.STAGING or self.race_status == RaceStatus.RACING:
+            self.stop()
+    
+        self.clear_laps()
+        self.race_status = RaceStatus.READY # Flag status as ready to start next race
+        self._racecontext.interface.set_race_status(RaceStatus.READY)
+        self.win_status = WinStatus.NONE
+        self.race_winner_name = ''
+        self.race_winner_phonetic = ''
+        self.race_winner_lap_id = 0
+        self.race_winner_pilot_id = RHUtils.PILOT_ID_NONE
+        self.race_leader_lap = 0  # clear current race leader
+        self.race_leader_pilot_id = RHUtils.PILOT_ID_NONE
+        self.status_message = ''
+        self._racecontext.rhui.emit_current_laps() # Race page, blank laps to the web client
+        self._racecontext.rhui.emit_current_leaderboard() # Race page, blank leaderboard to the web client
+        self._racecontext.rhui.emit_race_status() # Race page, to set race button states
+    
+        if 'saved' in kwargs and kwargs['saved'] == True:
+            # discarding follows a save action
+            pass
+        else:
+            # discarding does not follow a save action
+            self._racecontext.events.trigger(Evt.LAPS_DISCARD)
+            if self._racecontext.cluster:
+                self._racecontext.cluster.emitToSplits('discard_laps')
+    
+        self._racecontext.events.trigger(Evt.LAPS_CLEAR)
+    
+    def clear_laps(self):
+        '''Clear the current laps table.'''
+        self._racecontext.branch_race_obj()
+        self.reset_current_laps() # Clear out the current laps table
+        logger.info('Current laps cleared')
+
+    def reset_current_laps(self):
+        '''Resets database current laps to default.'''
+        self.node_laps = {}
+        for idx in range(self.num_nodes):
+            self.node_laps[idx] = []
+
+        self.clear_results()
+        logger.debug('Database current laps reset')
 
     def schedule(self, s, m=0):
         if self.race_status != RaceStatus.READY:
@@ -117,6 +987,107 @@ class RHRace():
 
     def check_all_nodes_finished(self):
         return False not in self.node_has_finished.values()
+
+    def check_win_condition(self, **kwargs):
+        previous_win_status = self.win_status
+        win_not_decl_flag = self.win_status in [WinStatus.NONE, WinStatus.PENDING_CROSSING, WinStatus.OVERTIME]
+        del_lap_flag = 'deletedLap' in kwargs
+    
+        # if winner not yet declared or racer lap was deleted then check win condition
+        win_status_dict = Results.check_win_condition_result(self._racecontext.race, self._racecontext.rhdata, self._racecontext.interface, **kwargs) \
+                          if win_not_decl_flag or del_lap_flag else None
+    
+        if win_status_dict is not None:
+            race_format = self.format
+            self.win_status = win_status_dict['status']
+    
+            if self.win_status != WinStatus.NONE and logger.getEffectiveLevel() <= logging.DEBUG:
+                logger.debug("Pilot lap counts: " + Results.get_pilot_lap_counts_str(self.results))
+                if race_format.team_racing_mode:
+                    logger.debug("Team lap totals: " + Results.get_team_lap_totals_str(self.team_results))
+    
+            # if racer lap was deleted and result is winner un-declared
+            if del_lap_flag and self.win_status != previous_win_status and \
+                                self.win_status == WinStatus.NONE:
+                self.win_status = WinStatus.NONE
+                self.race_winner_name = ''
+                self.race_winner_phonetic = ''
+                self.race_winner_lap_id = 0
+                self.race_winner_pilot_id = RHUtils.PILOT_ID_NONE
+                self.status_message = ''
+                logger.info("Race status msg:  <None>")
+                return win_status_dict
+    
+            if win_status_dict['status'] == WinStatus.DECLARED:
+                # announce winner
+                win_data = win_status_dict['data']
+                if race_format.team_racing_mode:
+                    win_str = win_data.get('name', '')
+                    team_win_str = self._racecontext.__('Team') + ' ' + win_str
+                    self.race_winner_name = team_win_str
+                    self.race_winner_phonetic = team_win_str
+                    status_msg_str = self._racecontext.__('Winner is') + ' ' + team_win_str
+                    log_msg_str = "Race status msg:  Winner is Team " + win_str
+                    phonetic_str = status_msg_str
+                else:
+                    win_str = win_data.get('callsign', '')
+                    self.race_winner_name = win_str
+                    self.race_winner_lap_id = win_data.get('laps', 0)
+                    status_msg_str = self._racecontext.__('Winner is') + ' ' + win_str
+                    log_msg_str = "Race status msg:  Winner is " + win_str
+                    win_pilot_id = win_data.get('pilot_id')
+                    if win_pilot_id:
+                        self.race_winner_pilot_id = win_pilot_id
+                        win_phon_name = self._racecontext.rhdata.get_pilot(win_pilot_id).phonetic
+                    elif win_data['callsign']:
+                        win_phon_name = win_data['callsign']
+                    else:
+                        win_phon_name = None
+                    if (not win_phon_name) or len(win_phon_name) <= 0:  # if no phonetic then use callsign
+                        win_phon_name = win_data.get('callsign', '')
+                    self.race_winner_phonetic = win_phon_name
+                    phonetic_str = self._racecontext.__('Winner is') + ' ' + win_phon_name
+    
+                # if racer lap was deleted then only output if win-status details changed
+                if (not del_lap_flag) or self.win_status != previous_win_status or \
+                                            status_msg_str != self.status_message:
+                    self.status_message = status_msg_str
+                    logger.info(log_msg_str)
+                    self._racecontext.rhui.emit_phonetic_text(phonetic_str, 'race_winner', True)
+                    self._racecontext.events.trigger(Evt.RACE_WIN, {
+                        'win_status': win_status_dict,
+                        'message': self.status_message,
+                        'node_index': win_data.get('node', None),
+                        'color': self.seat_colors[win_data['node']] \
+                                                if 'node' in win_data else None,
+                        })
+    
+            elif win_status_dict['status'] == WinStatus.TIE:
+                # announce tied
+                if win_status_dict['status'] != previous_win_status:
+                    self.status_message = self._racecontext.__('Race Tied')
+                    logger.info("Race status msg:  Race Tied")
+                    self._racecontext.rhui.emit_phonetic_text(self.status_message, 'race_winner')
+            elif win_status_dict['status'] == WinStatus.OVERTIME:
+                # announce overtime
+                if win_status_dict['status'] != previous_win_status:
+                    self.status_message = self._racecontext.__('Race Tied: Overtime')
+                    logger.info("Race status msg:  Race Tied: Overtime")
+                    self._racecontext.rhui.emit_phonetic_text(self.status_message, 'race_winner')
+    
+            if 'max_consideration' in win_status_dict:
+                logger.info("Waiting {0}ms to declare winner.".format(win_status_dict['max_consideration']))
+                gevent.sleep(win_status_dict['max_consideration'] / 1000)
+                if 'start_token' in kwargs and self.start_token == kwargs['start_token']:
+                    logger.info("Maximum win condition consideration time has expired.")
+                    self.check_win_condition(forced=True)
+    
+            if 'emit_leaderboard_on_win' in kwargs:
+                if self.win_status != WinStatus.NONE:
+                    self._racecontext.rhui.emit_current_leaderboard()  # show current race status on leaderboard
+        
+        return win_status_dict
+
 
     def get_active_laps(self, late_lap_flag=False):
         # return active (non-deleted) laps objects
